solr_search: automatic adjusting of maxRows
There is a performance penalty when asking for too many rows.
https://wiki.apache.org/solr/SolrPerformanceProblems#Asking_for_too_many_rows

If number of rows is specified and is higher than configured limit (50000),
solr_search first queries Solr for number of records with rows=0.
Then initial number of rows is replaced by real number of records
and solr_search is executed.